# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.BaseResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/dto.ErrorInfo'
      message:
        example: Operation completed successfully
        type: string
      success:
        example: true
        type: boolean
      timestamp:
        example: "2023-01-01T12:00:00Z"
        type: string
    type: object
...
  dto.ErrorInfo:
    properties:
      code:
        example: VALIDATION_ERROR
        type: string
      details:
        example: Name field is required
        type: string
      message:
        example: Invalid input parameters
        type: string
    type: object
...
  dto.PaginatedResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/dto.ErrorInfo'
      message:
        example: Operation completed successfully
        type: string
      pagination:
        $ref: '#/definitions/dto.PaginationInfo'
      success:
        example: true
        type: boolean
      timestamp:
        example: "2023-01-01T12:00:00Z"
        type: string
    type: object
...
host: localhost:9090
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A comprehensive vehicle spare parts shop management system with single-location
    inventory tracking, brand management, vehicle compatibility, unified purchase
    receipt processing, customer management, JWT authentication, and role-based access
    control. Features include product catalogs, vehicle-part compatibility, purchase
    order/goods receipt workflow, customer relationship management, audit trails,
    and role-based access control across the entire system.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Vehicle Spare Parts Shop Management API
  version: "1.0"
paths:
...
  /vehicle-models:
    get:
      consumes:
      - application/json
      description: Get a paginated list of vehicle models with optional filtering
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Search by name or code
        in: query
        name: search
        type: string
      - description: Filter by vehicle brand ID
        format: uuid
        in: query
        name: vehicle_brand_id
        type: string
      - description: Filter by active status
        in: query
        name: is_active
        type: boolean
      - description: Filter by year from
        in: query
        name: year_from
        type: integer
      - description: Filter by year to
        in: query
        name: year_to
        type: integer
      - description: Filter by fuel type
        in: query
        name: fuel_type
        type: string
      - description: Filter by transmission
        in: query
        name: transmission
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.VehicleModelDetailResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.BaseResponse'
      summary: List vehicle models
      tags:
      - Vehicle Models
    post:
      consumes:
      - application/json
      description: Create a new vehicle model with the provided information
      parameters:
      - description: Vehicle model creation request
        in: body
        name: vehicle_model
        required: true
        schema:
          $ref: '#/definitions/dto.CreateVehicleModelRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/dto.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.VehicleModelDetailResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.BaseResponse'
      summary: Create a new vehicle model
      tags:
      - Vehicle Models
  /vehicle-models/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a vehicle model by their ID (soft delete)
      parameters:
      - description: Vehicle Model ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.BaseResponse'
      summary: Delete a vehicle model
      tags:
      - Vehicle Models
    get:
      consumes:
      - application/json
      description: Get a specific vehicle model by their ID
      parameters:
      - description: Vehicle Model ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.VehicleModelDetailResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.BaseResponse'
      summary: Get vehicle model by ID
      tags:
      - Vehicle Models
    put:
      consumes:
      - application/json
      description: Update an existing vehicle model's information
      parameters:
      - description: Vehicle Model ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Vehicle model update request
        in: body
        name: vehicle_model
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateVehicleModelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.VehicleModelDetailResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.BaseResponse'
      summary: Update a vehicle model
      tags:
      - Vehicle Models
  /vehicle-models/{id}/activate:
    post:
      consumes:
      - application/json
      description: Activate a deactivated vehicle model
      parameters:
      - description: Vehicle Model ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.VehicleModelDetailResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.BaseResponse'
      summary: Activate a vehicle model
      tags:
      - Vehicle Models
  /vehicle-models/{id}/deactivate:
    post:
      consumes:
      - application/json
      description: Deactivate an active vehicle model
      parameters:
      - description: Vehicle Model ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.VehicleModelDetailResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.BaseResponse'
      summary: Deactivate a vehicle model
      tags:
      - Vehicle Models
  /vehicle-models/active:
    get:
      consumes:
      - application/json
      description: Get all active vehicle models
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.BaseResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.VehicleModelResponse'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.BaseResponse'
      summary: List active vehicle models
      tags:
      - Vehicle Models
  /vehicle-models/brand/{brand_id}:
    get:
      consumes:
      - application/json
      description: Get all vehicle models for a specific vehicle brand
      parameters:
      - description: Vehicle Brand ID
        format: uuid
        in: path
        name: brand_id
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.VehicleModelResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.BaseResponse'
      summary: Get vehicle models by brand ID
      tags:
      - Vehicle Models
  /vehicle-models/code/{code}:
    get:
      consumes:
      - application/json
      description: Get a specific vehicle model by their code
      parameters:
      - description: Vehicle Model code
        example: '"TOYT-CAMR01"'
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.VehicleModelDetailResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.BaseResponse'
      summary: Get vehicle model by code
      tags:
      - Vehicle Models
...
securityDefinitions:
  ApiKeyAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"